Feat: new dependencies, files and features
It was  installed new dependencies for redux/toolkit.  In the file index.js was implemented new redux feature. It was created new file named store.js for redux store. Also was created new file named user.reducer.js. It was added new feature in the file utils.services.js. In the files login.jsx and register.jsx was implemented the redux features